Update README.md
This is the next step where you create a file which has the computed values of each word. I think I have to use FileReader to reassign the integer values to each character, but I will reassess in the morning. THIS NEW EDITION IS NOT COMPLETE!!! I AM JUST SAVING WORK.